% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dashboard.R
\name{simple_dashboard}
\alias{simple_dashboard}
\title{Compute data for a simple seasonal adjustment}
\usage{
simple_dashboard(x, digits = 2, add_obs_to_forecast = TRUE)
}
\arguments{
\item{x}{A seasonal adjustment model.}

\item{digits}{Number of digits used in the tables.}

\item{add_obs_to_forecast}{Boolean indicating if the last observed values should be added to the forecast table (for the plot).}
}
\description{
Function to compute the data to produce a simple seasonal adjustment dashboard.
}
\examples{
data <- window(rjd3toolkit::ABS$X0.2.09.10.M, start = 2003)
sa_model <- rjd3x13::x13(data)
dashboard_data <- simple_dashboard(sa_model)
plot(dashboard_data, main = "Simple dashboard")
}
